Change $semantic-root-element to #__nuxt
pico.css apply padding for header, main and footer elements

that are direct children of `<body>`. Since nuxt app need

a root div to mount the application (`<div id="__nuxt">`),

the header, main and footer elements are not direct children

of the body tag.

To fix this we can set the variable `$semantic-root-element`

that will be used in pico.css instead of the body.

Now pico.css styling is aplied to header, main and footer

elements that are direct children of `<div id="__nuxt">`